pyproject.toml
[project]
name = "jfk-ocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu>=1.10.0",
    "langchain>=0.3.21",
    "langchain-community>=0.3.20",
    "langchain-openai>=0.3.9",
    "nltk>=3.9.1",
    "numpy>=2.2.4",
    "openai>=1.68.2",
    "rapidfuzz>=3.12.2",
    "reductoai>=0.3.0",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.2",
    "tiktoken>=0.9.0",
]
